- The builder has gotten fairly complex, so moving it to it's own file
